[FIX] 정산 관련 서비스 분리
