+++ libs/nest-modules/src/lib/seller/seller.controller.ts
@@ -23,6 +23,7 @@ import {
 import { JwtAuthGuard } from '../_nest-units/guards/jwt-auth.guard';
 import { MailVerificationService } from '../auth/mailVerification.service';
 import { SellerService } from './seller.service';
+import { SellerSettlementService } from './seller-settlement.service';
 import { SellerInfo } from '../_nest-units/decorators/sellerInfo.decorator';
 import { UserPayload } from '../auth/auth.interface';
 
@@ -30,6 +31,7 @@ import { UserPayload } from '../auth/auth.interface';
 export class SellerController {
   constructor(
     private readonly sellerService: SellerService,
+    private readonly sellerSettlementService: SellerSettlementService,
     private readonly mailVerificationService: MailVerificationService,
   ) {}
 
@@ -91,7 +93,7 @@ export class SellerController {
   @UseGuards(JwtAuthGuard)
   @Get('settlement')
   public async selectSellerSettlementInfo(@SellerInfo() sellerInfo: UserPayload) {
-    return this.sellerService.selectSellerSettlementInfo(sellerInfo);
+    return this.sellerSettlementService.selectSellerSettlementInfo(sellerInfo);
   }
 
   // 본인의 사업자 등록증 등록
@@ -101,7 +103,7 @@ export class SellerController {
     @Body(ValidationPipe) dto: BusinessRegistrationDto,
     @SellerInfo() sellerInfo: UserPayload,
   ) {
-    return this.sellerService.insertBusinessRegistration(dto, sellerInfo);
+    return this.sellerSettlementService.insertBusinessRegistration(dto, sellerInfo);
   }
 
   // 본인의 사업자 등록증 등록
@@ -111,6 +113,6 @@ export class SellerController {
     @Body(ValidationPipe) dto: SettlementAccountDto,
     @SellerInfo() sellerInfo: UserPayload,
   ) {
-    return this.sellerService.insertSettlementAccount(dto, sellerInfo);
+    return this.sellerSettlementService.insertSettlementAccount(dto, sellerInfo);
   }
 }
